INITIAL COMMIT
First data generation file creating a vector xvec of length 1000
between 0 and 100 and computes a vector yvec from it according to
the linear relationship y = 0.5x-2. Then, a random noise vector is
created with uniformly random values between 0 and 10. These noise
values are added to the yvec vector resulting in yvec_hat.

The vectors xvec and yvec_hat are written to a file.

The data in the resulting file can be used as an ultra-basic example
data set for linear regression tasks.